refactor: move testContext to setup_test.go
Move testContext struct and newTestContext function from steps_test.go to setup_test.go for better code organization.
